Rename coroc module
