+++ compiler/color.go
