+++ compiler/testdata/coroutine.go
